Add a PropertiesFileTransformer as suggested in #73
This is a verbatim copy of the class found in the Griffon build.
Please consider these files a starting point for fullfiling the
feature described in #73. The current implementation is good
enough for the uses cases found in Griffon projects, it's likely
to be broad enough to cover all cases, bu one can never tell ;-)

The license governing the shadow-gradle-plugin project has been
applied to the new files :-)