Introduce keyTransformer property for PropertiesFileTransformer
The new keyTransformer property can be used to transform keys in
properties files, e.g. because they contain class names about to be
relocated. Its value can be set to a closure that receives the original
key and returns the key name to be used in the resulting properties
file.