Implemented JAR minimization
This commit adds two new configuration options to the 'shadowJar' task.
When 'minimize' is set the shadow JAR would only include the classes
for the project the task operates on and its dependencies. The user can
protect other classes from minimization by exluding their module.

Here's an example configuration

    shadowJar {
      minimize {
        exclude(dependency('junit:junit:.*'))
      }
    }

This commit is a cherry-pick of 825814b from '2.0.1' branch.

Change-Id: I0302d801c4cc5a63ad2daba07164c2de34e159c9
JIRA: RP-3087