Add support for Gradle metadata
This commit adds support for Gradle metadata, available in
Gradle 5.3+. First, it makes use of the new public publication
API to create a custom software component.

This software component is registered so that when published,
we publish both the normal jars _and_ the shadow jar. When
Gradle metadata is enabled, we also publish Gradle metadata,
which will declare variants for all usages. In particular,
it makes use of the `Bundling` attribute to tell Gradle that
this particular variant is a shadow jar.

This makes it possible for consumers to _choose_ whatever
dependency they want: either the regular jar, or the shadow
one. It's worth noting that for Maven consumers, they would
have to rely on the classifier to get the all jar.