Fix deprecation when running on Gradle 8.8
`CopyProcessingSpec.fileMode` was recently deprecated. It is replaced by
the `filePermissions` method, that was introduced in 8.3.
For that reason this change updates the build to 8.3 to be able to call
`filePermissions`. The code then uses that method if running on 8.3 or
higher. If not, it falls back to calling `fileMode`.