Allow using file trees of JARs together with the configuration cache
We are adding JARs from a directory that is built by a task to the `ShadowJar` by using `includedDependencies`. That doesn't seem to work when using the configuration cache, since the contents of the directory are evaluated at configuration time, causing them always to be empty.

This PR fixes that by doing the actual resolution as late as in the task action, not when the `includedZipTrees` provider is being stored in the configuration cache.