Don't re-add suppressed Gradle API to compileOnly configuration (#1422)
As part of Gradle's experiments with publishing an external Gradle API (gradle/gradle issue 29483), they are beginning to look at suppressing the local Gradle API dependency that is usually used when the `java-gradle-plugin` plugin is applied. Because of this, though, the current strategy that the Shadow plugin uses to move it from `api` to `compileOnly` doesn't check if it exists to begin with. This PR aims to solve that problem. I've added tests as needed.

Gradle's Plugin Publish plugin has a similar issue which I've gone into detail in gradle/plugin-portal-requests issue 260.


* Add tests for checking presence of Gradle API in dependencies

* Clean up tests

* Less diff in configureJavaGradlePlugin

* Remove afterEvaluate

* Apply suggestions from code review

* Update changelog

* Try `named`

* Revert "Try `named`"

This reverts commit 0cae38effa42a55462bb8fe0c9673e9e840517ae.

---------

Co-authored-by: Goooler <wangzongler@gmail.com>