Use BufferedOutputStream when writing the Zip file to improve performance (#1579)
* fix: use buffered writes for performance

When not using STORED entries use a BufferedOutputStream to avoid lots
of small writes to the file system.

Testing this with a 300mb jar build I see the total build time going
from 40s to 30s.

Note that it is not possible to do this with STORED entries as the
implementation requires a RandomAccessFile to update the CRC after
write.

* Cleanup

---------

Co-authored-by: Goooler <wangzongler@gmail.com>