Make assemble depend on shadowJar even if it is added later (#1766)
Currently, Shadow makes the assemble task depend on shadowJar if and only if it exists when the shadow plugin is applied. This is usually fine, but in the case of some unorthodox setups, it can mean that if the Shadow plugin is applied before the base plugin, the assemble task can end up existing without depending on the shadowJar task.

To address this, instead of using `TaskContainer#findByName(String)`, we can use `#named(Spec<String>)` to work with a live view of the container. This is very similar to how `#withType(Class<S extends T>` works, where `configureEach` applies the action to each member that passes the spec, even if it is added later.

---------

Co-authored-by: Goooler <wangzongler@gmail.com>