Make the output of PropertiesFileTransformer reproducible (#1861)
* `PropertiesFileTransformer` - make merged properties reproducible

`PropertiesFileTransformer` leverages `java.util.Properties`, which relies on `java.util.Hashtable`. The serialized properties are not guaranteed to be reproducible.

This change changes the transformer to use a sorted map to generate reproducible output. The existing class `CleanProperties` is replaced with `ReproducibleProperties`.
Functionality around charset handling is retained, and extended with a functionality to generate unicode escapes (ASCII output).

* Cleanups

---------

Co-authored-by: Goooler <wangzongler@gmail.com>